feat(settings): store widget position as distance in pixels from the closest side of the screen
This basically mimicks how positioning is done via Movable.

This ensures the widget is placed at the exact same place in relation to the edge of the screen, regardless of the size of the window.

Anchoring this way makes sense since most people just place the thermometer to the side, leaving the center empty to see the streetview panorama.

The switch to absolute pixels instead of percentages also enforces consistency when resizing the window.